updated health check
