+++ src/app/Constants.cs
@@ -34,5 +34,7 @@
 
         public const int DefaultPageSize = 100;
         public const int MaxPageSize = 1000;
+
+        public const string CosmosHealthCheck = "CosmosHealthCheck";
     }
 }
