+++ src/app/HealthChecks/CosmosHealthCheckBuilderExtensions.cs
@@ -1,0 +1,35 @@
+using Helium.DataAccessLayer;
+using Helium.Model;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Azure.Cosmos;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Threading;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Microsoft.Extensions.Options;
+using Microsoft.AspNetCore.Http;
+using Newtonsoft.Json;
+
+namespace Helium
+{
+    public static class CosmosHealthCheckBuilderExtensions
+    {
+        public static IHealthChecksBuilder AddCosmosHealthCheck(
+            this IHealthChecksBuilder builder,
+            string name,
+            HealthStatus? failureStatus = null,
+            IEnumerable<string> tags = null)
+        {
+            // Register a check of type Cosmos
+            builder.AddCheck<CosmosHealthCheck>(name, failureStatus ?? HealthStatus.Degraded, tags);
+
+            return builder;
+        }
+    }
+}
